requirements.txt
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.23.0
pillow==9.3.0
einops==0.6.1
ftfy==6.1.1
fvcore==0.1.5
regex==2023.8.8
rsa==4.9
scipy==1.11.2
sympy==1.11.1
timm==0.9.7
torchelastic==0.2.2
torchsummary==1.5.1
yacs==0.1.8
pytorchvideo==0.1.5
hydra-core==1.3.2
wilds==1.2.2
gdown==4.7.1
tabulate==0.9.0
huggingface-hub==0.18.0